+++ ThreeDSecure/src/main/java/com/braintreepayments/api/threedsecure/ThreeDSecureClient.kt
@@ -52,8 +52,8 @@ class ThreeDSecureClient internal constructor(
      *
      * Verification is associated with a transaction amount and your merchant account. To specify a
      * different merchant account (or, in turn, currency), you will need to specify the merchant
-     * account id when [
-     * generating a client token](https://developer.paypal.com/braintree/docs/start/hello-client#get-a-client-token)
+     * account id when
+     * [generating a client token](https://developer.paypal.com/braintree/docs/start/hello-client#get-a-client-token).
      *
      * During lookup the original payment method nonce is consumed and a new one is returned, which
      * points to the original payment method, as well as the 3D Secure verification. Transactions
@@ -77,7 +77,24 @@ class ThreeDSecureClient internal constructor(
         }
     }
 
-    private suspend fun createPaymentAuthRequest(
+    /**
+     * Call this method to initiate the 3D Secure flow.
+     *
+     * Verification is associated with a transaction amount and your merchant account. To specify a
+     * different merchant account (or, in turn, currency), you will need to specify the merchant
+     * account id when
+     * [generating a client token](https://developer.paypal.com/braintree/docs/start/hello-client#get-a-client-token).
+     *
+     * During lookup the original payment method nonce is consumed and a new one is returned, which
+     * points to the original payment method, as well as the 3D Secure verification. Transactions
+     * created with this nonce will be 3D Secure, and benefit from the appropriate liability shift
+     * if authentication is successful or fail with a 3D Secure failure.
+     *
+     * @param context  Android context
+     * @param request  the [ThreeDSecureRequest] with information used for authentication.
+     * @return [ThreeDSecurePaymentAuthRequest]
+     */
+    suspend fun createPaymentAuthRequest(
         context: Context,
         request: ThreeDSecureRequest
     ): ThreeDSecurePaymentAuthRequest {
@@ -167,7 +184,15 @@ class ThreeDSecureClient internal constructor(
         }
     }
 
-    private suspend fun prepareLookup(
+    /**
+     * Creates a stringified JSON object containing the information necessary to perform a lookup
+     *
+     * @param context  Android Context
+     * @param request  the [ThreeDSecureRequest] that has a nonce and an optional UI
+     * customization.
+     * @return [ThreeDSecurePrepareLookupResult]
+     */
+    suspend fun prepareLookup(
         context: Context,
         request: ThreeDSecureRequest
     ): ThreeDSecurePrepareLookupResult {
@@ -183,7 +208,7 @@ class ThreeDSecureClient internal constructor(
                         .put("requestedThreeDSecureVersion", "2")
                         .put("sdkVersion", "Android/${BuildConfig.VERSION_NAME}")
                 )
-        } catch (ignored: JSONException) {
+        } catch (_: JSONException) {
         }
 
         try {
@@ -226,22 +251,36 @@ class ThreeDSecureClient internal constructor(
      *
      * @param lookupResponse The lookup response from the server side call to lookup the 3D Secure
      * information.
-     * @param callback       [ThreeDSecureResultCallback]
+     * @param callback [ThreeDSecureResultCallback]
      */
     fun initializeChallengeWithLookupResponse(
         lookupResponse: String,
         callback: ThreeDSecurePaymentAuthRequestCallback
     ) {
         coroutineScope.launch {
-            val result = try {
-                braintreeClient.getConfiguration()
-                sendAnalyticsAndResult(fromJson(lookupResponse))
-            } catch (e: Exception) {
-                if (e is CancellationException) throw e
-                braintreeClient.sendAnalyticsEvent(ThreeDSecureAnalytics.LOOKUP_FAILED)
-                createPaymentAuthFailure(e)
-            }
-            callback.onThreeDSecurePaymentAuthRequest(result)
+            callback.onThreeDSecurePaymentAuthRequest(
+                initializeChallengeWithLookupResponse(lookupResponse)
+            )
+        }
+    }
+
+    /**
+     * Initialize a challenge from a server side lookup call.
+     *
+     * @param lookupResponse The lookup response from the server side call to lookup the 3D Secure
+     * information.
+     * @return [ThreeDSecurePaymentAuthRequest]
+     */
+    suspend fun initializeChallengeWithLookupResponse(
+        lookupResponse: String
+    ): ThreeDSecurePaymentAuthRequest {
+        return try {
+            braintreeClient.getConfiguration()
+            sendAnalyticsAndResult(fromJson(lookupResponse))
+        } catch (e: Exception) {
+            if (e is CancellationException) throw e
+            braintreeClient.sendAnalyticsEvent(ThreeDSecureAnalytics.LOOKUP_FAILED)
+            createPaymentAuthFailure(e)
         }
     }
 
@@ -284,8 +323,8 @@ class ThreeDSecureClient internal constructor(
     }
 
     /**
-     * Call this method from the [ThreeDSecureLauncherCallback] passed to the
-     * [ThreeDSecureLauncher] used to launch the 3DS authentication challenge.
+     * When launching the 3DS challenge, call this method from the [ThreeDSecureLauncherCallback]
+     * passed to the [ThreeDSecureLauncher].
      *
      * @param paymentAuthResult a [ThreeDSecurePaymentAuthResult] received in [ThreeDSecureLauncherCallback]
      * @param callback       a [ThreeDSecureResultCallback]
@@ -300,8 +339,15 @@ class ThreeDSecureClient internal constructor(
         }
     }
 
+    /**
+     * When launching the 3DS challenge, call this method from the [ThreeDSecureLauncherCallback]
+     * passed to the [ThreeDSecureLauncher].
+     *
+     * @param paymentAuthResult a [ThreeDSecurePaymentAuthResult] received in [ThreeDSecureLauncherCallback]
+     * @return [ThreeDSecureResult]
+     */
     @Suppress("LongMethod")
-    private suspend fun tokenize(
+    suspend fun tokenize(
         paymentAuthResult: ThreeDSecurePaymentAuthResult
     ): ThreeDSecureResult {
         val threeDSecureError = paymentAuthResult.error